refactor int

+++ lab_09_01_1/calculating_matrix.c
@@ -2,21 +2,21 @@
 #include "resizing_matrix.h"
 #include "user_interface.h"
 
-int64_t calc_mean_column(matrix a_m, int64_t col)
+int calc_mean_column(matrix a_m, int col)
 {
     double mean = 0;
-    for (int64_t i = col; i < a_m.n * a_m.m - a_m.m; i += a_m.m)
+    for (int i = col; i < a_m.n * a_m.m - a_m.m; i += a_m.m)
     {
         mean += *(a_m.p + i);
     }
     mean /= AN;
-    return (int64_t)floor(mean);
+    return (int)floor(mean);
 }
 
-int64_t find_max_row(matrix a_m, int64_t row)
+int find_max_row(matrix a_m, int row)
 {
-    int64_t max_el = *((a_m.p) + row * a_m.m);
-    for (int64_t *n = (a_m.p) + row * a_m.m; n < a_m.p + row * a_m.m + AM; n++)
+    int max_el = *((a_m.p) + row * a_m.m);
+    for (int *n = (a_m.p) + row * a_m.m; n < a_m.p + row * a_m.m + AM; n++)
     {
         if ((max_el) < *n)
         {
@@ -26,12 +26,12 @@ int64_t find_max_row(matrix a_m, int64_t row)
     return max_el;
 }
 
-int64_t find_max(matrix a_m, int64_t *maxn, int64_t *maxm)
+int find_max(matrix a_m, int *maxn, int *maxm)
 {
-    int64_t max = *a_m.p;
-    for (int64_t i = 0; i < a_m.n; i++)
+    int max = *a_m.p;
+    for (int i = 0; i < a_m.n; i++)
     {
-        for (int64_t j = 0; j < a_m.m; j++)
+        for (int j = 0; j < a_m.m; j++)
         {
             if (CUR_ELEMENT > max)
             {
@@ -44,23 +44,23 @@ int64_t find_max(matrix a_m, int64_t *maxn, int64_t *maxm)
     return OK;
 }
 
-int64_t multiply_matrix(matrix a_m, matrix b_m, matrix *m_m)
+int multiply_matrix(matrix a_m, matrix b_m, matrix *m_m)
 {
-    for (int64_t i = 0; i < a_m.n; i++)
+    for (int i = 0; i < a_m.n; i++)
     {
-        for (int64_t j = 0; j < a_m.m; j++)
+        for (int j = 0; j < a_m.m; j++)
         {
             *((m_m->p) + j + i * (a_m.m)) = 0;
-            for (int64_t k = 0; k < a_m.m; k++)
+            for (int k = 0; k < a_m.m; k++)
                 *((m_m->p) + j + i * (a_m.m)) += *((a_m.p) + k + i * (a_m.m)) * *((b_m.p) + j + k * (b_m.m));
         }
     }
     return OK;
 }
 
-int64_t copy_matrix(matrix old, matrix *newm)
+int copy_matrix(matrix old, matrix *newm)
 {
-    for (int64_t i = 0; i < newm->m * newm->n; i++)
+    for (int i = 0; i < newm->m * newm->n; i++)
     {
         *(newm->p + i) = *(old.p + i);
     }
@@ -69,18 +69,18 @@ int64_t copy_matrix(matrix old, matrix *newm)
 
 void create_ematrix(matrix *e_m)
 {
-    for (int64_t i = 0; i < e_m->m * e_m->n; i++)
+    for (int i = 0; i < e_m->m * e_m->n; i++)
     {
         *(e_m->p + i) = 0;
     }
 
-    for (int64_t i = 0; i < e_m->m * e_m->n; i += e_m->m + 1)
+    for (int i = 0; i < e_m->m * e_m->n; i += e_m->m + 1)
     {
         *(e_m->p + i) = 1;
     }
 }
 
-int64_t expo_matrix(matrix *a_m, int64_t power)
+int expo_matrix(matrix *a_m, int power)
 {
     matrix ap_m;
     ap_m.m = a_m->m;
@@ -112,7 +112,7 @@ int64_t expo_matrix(matrix *a_m, int64_t power)
         create_ematrix(&r_m);
     }
 
-    for (int64_t i = 0; i < power - 1; i++)
+    for (int i = 0; i < power - 1; i++)
     {
         multiply_matrix(*a_m, ap_m, &r_m);
         copy_matrix(r_m, &ap_m);
